sof: crossover: Fix iir_df2t utilization bug
In Seb's implementation, lr4->coef has been reused by the biquad series
because they are always identical. For iir_df2t function coefficients are
assigned per biquad in a row. This is the root cause we've got incorrect
filter outputs recently.

Note that this commit is temporary for demonstration and should be merged
to the previous commit.

Signed-off-by: Pin-chih Lin <johnylin@google.com>